moved *_EVENT constants into ./constants/index.js
