+++ src/js/constants/index.js
@@ -1,0 +1,8 @@
+/* @flow */
+'use strict';
+
+export const CORE_EVENT: 'CORE_EVENT' = 'CORE_EVENT';
+export const UI_EVENT: 'UI_EVENT' = 'UI_EVENT';
+export const DEVICE_EVENT: 'DEVICE_EVENT' = 'DEVICE_EVENT';
+export const TRANSPORT_EVENT: 'TRANSPORT_EVENT' = 'TRANSPORT_EVENT';
+export const RESPONSE_EVENT: 'RESPONSE_EVENT' = 'RESPONSE_EVENT';
