+++ src/js/core/CoreMessage.js
@@ -1,13 +1,9 @@
 /* @flow */
 'use strict';
 
+import { UI_EVENT, DEVICE_EVENT, TRANSPORT_EVENT, RESPONSE_EVENT } from '../constants';
 import type { CoreMessage } from 'flowtype';
 import type { UiMessageFactory } from 'flowtype/ui-message';
-
-export const UI_EVENT: 'UI_EVENT' = 'UI_EVENT';
-export const DEVICE_EVENT: 'DEVICE_EVENT' = 'DEVICE_EVENT';
-export const TRANSPORT_EVENT: 'TRANSPORT_EVENT' = 'TRANSPORT_EVENT';
-export const RESPONSE_EVENT: 'RESPONSE_EVENT' = 'RESPONSE_EVENT';
 
 // parse MessageEvent .data object into CoreMessage
 export const parseMessage = (messageData: any): CoreMessage => {
